Main

+++ Unigram/Unigram/Common/WindowContext.cs
@@ -240,7 +240,7 @@ namespace Unigram.Common
                 var query = "tg://";
 
                 var contactId = await ContactsService.GetContactIdAsync(share.ShareOperation.Contacts.FirstOrDefault());
-                if (contactId is int userId)
+                if (contactId is long userId)
                 {
                     var response = await _lifetime.ActiveItem.ProtoService.SendAsync(new CreatePrivateChat(userId, false));
                     if (response is Chat chat)
@@ -302,7 +302,7 @@ namespace Unigram.Common
                 }
 
                 var contactId = await ContactsService.GetContactIdAsync(contact.Contact.Id);
-                if (contactId is int userId)
+                if (contactId is long userId)
                 {
                     var response = await _lifetime.ActiveItem.ProtoService.SendAsync(new CreatePrivateChat(userId, false));
                     if (response is Chat chat)
